services/ai-svc/requirements.txt
fastapi[standard]>=0.115,<1
uvicorn[standard]>=0.34,<1
pydantic>=2.10,<3
pydantic-settings>=2.8,<3
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.30,<1
redis[hiredis]>=5.2,<6
nats-py>=2.9,<3
litellm>=1.40,<2
langfuse>=2.0,<3
pgvector>=0.3,<1
pypdf>=4.0,<6
pyphen>=0.16,<1
httpx>=0.28,<1
cryptography>=43,<45
PyJWT[crypto]>=2.9,<3
